+++ src/libpcp/src/pdu.c
@@ -285,7 +285,7 @@ __pmXmitPDU(int fd, __pmPDU *pdubuf)
 	    *p++ = '~';	/* buffer end */
 
 	if (mypid == -1)
-	    mypid = getpid();
+	    mypid = (int)getpid();
 	fprintf(stderr, "[%d]pmXmitPDU: %s fd=%d len=%d",
 		mypid, __pmPDUTypeStr(php->type), fd, php->len);
 	for (j = 0; j < jend; j++) {
@@ -490,7 +490,7 @@ check_read_len:
 	    *p++ = '~';	/* buffer end */
 
 	if (mypid == -1)
-	    mypid = getpid();
+	    mypid = (int)getpid();
 	fprintf(stderr, "[%d]pmGetPDU: %s fd=%d len=%d from=%d",
 		mypid, __pmPDUTypeStr(php->type), fd, php->len, php->from);
 	for (j = 0; j < jend; j++) {
